requirements.txt
streamlit
requests
rembg
Pillow
numpy